getClimbLevel() now returns an int, not a double

+++ MISTAKE2019/src/Robot.java
@@ -48,7 +48,7 @@ public class Robot {
 		return;
 	}
 	
-	public double getClimbLevel() {
+	public int getClimbLevel() {
 		return climbLevel;
 	}
